refactored code; kept filenames consistent
