+++ server/models/User.js
@@ -20,25 +20,13 @@ const userSchema = new mongoose.Schema({
     minlength: [8, "Your password must have at least 8 characters"],
     select: false
   },
-  passwordConfirm: {
-    type: String,
-    required: [true, "Please confirm your password"],
-    validate: {
-      validator: function(val) {
-        return val === this.password;
-      },
-      message: `Passwords don't match`
-    }
-  },
   avatar: String
 });
 
 // hash password before saving to database
 userSchema.pre("save", async function(next) {
   if (!this.isModified("password")) return next();
-
   this.password = await bcrypt.hash(this.password, 12);
-  this.passwordConfirm = undefined;
   next();
 });
 
@@ -51,5 +39,4 @@ userSchema.methods.isPasswordCorrect = async function(
 };
 
 const User = mongoose.model("User", userSchema);
-
 export default User;
